refactor: первые шагив написании запроса на буккинг, поправлен запрос к поиску веще по userId
