+++ src/main/java/ru/practicum/shareit/booking/enums/BookingStatus.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.booking;
+package ru.practicum.shareit.booking.enums;
 
 /*
     Статус бронирования
