+++ src/main/java/ru/practicum/shareit/booking/mapper/BookingMapper.java
@@ -1,0 +1,4 @@
+package ru.practicum.shareit.booking.mapper;
+
+public class BookingMapper {
+}
