+++ src/main/java/ru/practicum/shareit/booking/storage/BookingRepository.java
@@ -1,0 +1,7 @@
+package ru.practicum.shareit.booking.storage;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import ru.practicum.shareit.booking.model.Booking;
+
+public interface BookingRepository extends JpaRepository<Booking, Long> {
+}
